projeto2_flask_sqlalchemy/app.py
from flask import Flask, request, jsonify
from flask_sqlalchemy import SQLAlchemy
from dotenv import load_dotenv
import os

# Load environment variables from .env file
load_dotenv()

app = Flask(__name__)

app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')

db = SQLAlchemy(app)

@app.route('/users', methods=['GET']) 
def hello():
    return jsonify({'message': 'Hello, World!'})  

if __name__ == '__main__':
    app.run(debug=True)