projeto2_flask_sqlalchemy/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==3.0.10
Werkzeug==2.3.8
flask-swagger-ui==4.11.1
pytest==7.4.3
python-dotenv